Improve Error Message in "Lab: Traffic Sign Inference
In the Transfer Learning lesson, the point of the "Lab: Traffic Sign Inference" concept is to teach students to resize images. And the text node does a good job of that.
But the error that actually appears when students run the code is unhelpful. It basically relates to an undefined variable, not to a mis-matched images sizes.
Let's improve the code so that it throws some sort of dimension mis-match exception when it's run.