scipy.misc.imread is gone, replace with matplotlib.pyplot.imread
In new 1.0.0 release of scipy (which is the current stable version as of Oct 25), `scipy.misc.imread` is gone (see the release notes: https://scipy.github.io/devdocs/release.1.0.0.html#deprecated-features).

I replaced it with `matplotlib.pyplot.imread` since `matplotlib` is already a dependency of several of the labs/projects. `pyplot`'s implementation has the quirk that it returns values in the range 0.0-1.0 instead of 0-255, so you need to scale the value by 255 before feeding it into the network.